apps/voice-agent/pyproject.toml
[project]
name = "gaia-voice-agent"
version = "0.1.0"
description = "Voice agent worker for GAIA"
requires-python = ">=3.11"

dependencies = [
  "gaia-shared",
  "aiohttp>=3.9.0",
  "livekit>=1.0.12",
  "livekit-api",
  "livekit-agents[turn-detector,silero,deepgram]~=1.2",
  "livekit-plugins-noise-cancellation~=0.2",
  "livekit-plugins-elevenlabs~=1.2",
  "hf_transfer",
  "python-dotenv",
]

[dependency-groups]
dev = [
  "ruff>=0.4.4",
]

[tool.uv.sources]
gaia-shared = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
